@@ -27,7 +27,7 @@ module Spree
         end
 
         respond_to do |format|
-          format.js { render text: 'Ok' }
+          format.js { render plain: 'Ok' }
         end
       end
 
