@@ -1,4 +1,4 @@
-class UpdateRelationTypes < ActiveRecord::Migration
+class UpdateRelationTypes < SpreeExtension::Migration[4.2]
   def up
     Spree::RelationType.where(applies_to: 'Product').update_all(applies_to: 'Spree::Product')
   end
