@@ -1,4 +1,4 @@
-class UpdateRelations < ActiveRecord::Migration
+class UpdateRelations < SpreeExtension::Migration[4.2]
   def up
     Spree::Relation.where(relatable_type: 'Product').update_all(relatable_type: 'Spree::Product')
     Spree::Relation.where(related_to_type: 'Product').update_all(related_to_type: 'Spree::Product')
